+++ src/models/index.ts
@@ -1,2 +1,1 @@
-export * from './QuestionnaireState'
-export * from './ValidResponse'
+export * from './SurveyState'