+++ CHANGELOG.md
@@ -4,5 +4,8 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # 0.0.1 (2023-07-20)
-
 - initial version
+
+# 0.1.0 (2024-05-07)
+- Adapt to Survey Structure to Handlers, Models, Messages, Repository and Services
+- Replace 'questionnaire' object names by survey